Add userTags to global data Remove tags from eleventyConfig. Add userTags as a global data object (tags is reserved in eleventy.) Add helpers/tags to get tagged discussions, subjects and collections by zooniverse ID.
