+++ src/api/_data/userTags.js
@@ -1,0 +1,7 @@
+const tags = require('../../helpers/tags');
+const userCollections = require('./userCollections');
+
+module.exports = async function() {
+  await userCollections();
+  return tags();
+}
