+++ src/api/templates/tag.njk
@@ -1,0 +1,9 @@
+---
+pagination:
+  data: userTags
+  resolve: values
+  size: 1
+  alias: tag
+permalink: "/userTags/{{ tag.name }}.json"
+---
+{{ tag | dump | safe }}
