+++ src/helpers/store.js
@@ -2,6 +2,7 @@ module.exports = {
   boards: {},
   discussions: {},
   subjects: {},
+  tags: {},
   users: {},
   userCollections: {}
 };
